+++ CHANGELOG.md
@@ -14,6 +14,9 @@ The main work (all changes without a GitHub username in brackets in the below li
 - @matter/node
     - Feature: We now allocate node IDs as sequential numbers; the old behavior of randomized node behavior is available if you set `ControllerBehavior` state property `nodeIdAssignment` to `"random"`
 
+- @project-chip/matter.js
+  - Fix: Prevent PairedNode from updating it's structure when node is already decommissioned
+
 ## 0.16.8 (2026-01-30)
 
 - @matter/node